requirements.txt
numpy==1.22.3
opencv-python==4.5.2.52
onnxruntime-gpu==1.10
pydantic==1.9.0